// src/logic/types.ts
export enum Element {
  Burned_Something = "burned something",
  Reed = "reed",
  Fish_Soup = "fish soup",
  Koumiss = "koumiss",
  Surstromming = "surströmming",
  Bush = "bush",
  Cotton_Candy = "cotton candy",
  Eggnog = "eggnog",
  Ice_Tea = "ice tea",
  Kvass = "kvass",
  Sushi = "sushi",
  Baklava = "baklava",
  Pilaf = "pilaf",
  Muesli = "muesli",
  Gozinaki = "gozinaki",
  Mead = "mead",
  Gingerbread = "gingerbread",
  Manure = "manure",
  Tree = "tree",
  Fondue = "fondue",
  Creme_Brulee = "creme brulee",
  Croissant = "croissant",
  Hot_Dog = "hot-dog",
  Hamburger = "hamburger",
  Popcorn = "popcorn",
  Communion = "communion",
  Beer = "beer",
  Vodka = "vodka",
  Brandy = "brandy",
  Liqueur = "liqueur",
  Cider = "cider",
  Soda = "soda",
  Cola = "cola",
  Mulled_Wine = "mulled wine",
  Ground_Coffee = "ground coffee",
  Cocoa_Powder = "cocoa powder",
  Cocoa_Bean = "cocoa bean",
  Coffee_Bean = "coffee bean",
  Shawerma = "shawerma",
  Animal = "animal",
  Earth = "earth",
  Furnace = "furnace",
  Ice = "ice",
  Tool = "tool",
  Bacteria = "bacteria",
  Egg = "egg",
  Meat = "meat",
  Milk = "milk",
  Water = "water",
  Boiling_Water = "boiling water",
  Omelette = "omelette",
  Kebab = "kebab",
  Forcemeat = "forcemeat",
  Kefir = "kefir",
  Cream = "cream",
  Cheese = "cheese",
  Blue_Cheese = "blue cheese",
  Boiled_Egg = "boiled egg",
  Roast = "roast",
  Condensed_Milk = "condensed milk",
  Milkshake = "milkshake",
  Fish = "fish",
  Seed = "seed",
  Grass = "grass",
  Sugar = "sugar",
  Sausage = "sausage",
  Tea = "tea",
  Cutlet = "cutlet",
  Toffee = "toffee",
  Ice_Cream = "ice cream",
  Fried_Fish = "fried fish",
  Cereal = "cereal",
  Mojito = "mojito",
  Spices = "spices",
  Flower = "flower",
  Caramel = "caramel",
  Syrup = "syrup",
  Butter_Dough = "butter dough",
  Flour = "flour",
  Porridge = "porridge",
  Berry = "berry",
  Sauce = "sauce",
  Nut = "nut",
  Honey = "honey",
  Candy = "candy",
  Cookie = "cookie",
  Pie = "pie",
  Cake = "cake",
  Dough = "dough",
  Juice = "juice",
  Wine = "wine",
  Jam = "jam",
  Yogurt = "yogurt",
  Marshmallow = "marshmallow",
  Mayonnaise = "mayonnaise",
  Fruit = "fruit",
  Nut_Butter = "nut butter",
  Mustard = "mustard",
  Topping = "topping",
  Bread = "bread",
  Pita = "pita",
  Baozi = "baozi",
  Spaghetti = "spaghetti",
  Noodles = "noodles",
  Alcohol = "alcohol",
  Cocoa = "cocoa",
  Chocolate = "chocolate",
  Coffee = "coffee",
  Latte = "latte",
  Vegetable = "vegetable",
  Toast = "toast",
  Sandwich = "sandwich",
  Hot_Chocolate = "hot chocolate",
  Eskimo = "eskimo",
  Chocolate_Bar = "chocolate bar",
  Muffin = "muffin",
  French_Fries = "french fries",
  Salad = "salad",
  Soup = "soup",
  Ketchup = "ketchup",
}

export enum FoodType {
  Animal = "animal",
  Bakery = "bakery",
  Drink = "drink",
  Meat = "meat",
  Milk = "milk",
  Other = "other",
  Plant = "plant",
  Inedible = "inedible",
  Fastfood = "fastfood",
  Vegan = "vegan",
  Sweet = "sweet",
}

export type ElementEntriesType = [Element, FoodType];

export type RecipesType = [[Element, Element], Element | Element[]][];

export type RecipesByElementType = {
  [key in Element]?: {
    [key in Element]?: Element | Element[];
  };
};